requirements.txt
networkx
pyvis
python-dotenv